re-organizing assets and creating a new jump
