+++ Assets/Scripts/PowerUp/PowerUps/Jump.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a6ed46b89c0e2c4aab2c65554100fd9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
