+++ Assets/Scripts/PowerUp/PowerUps/Jump/DoubleJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e039bb888ea4a84f9e131b8665a9882
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
