+++ Assets/Scripts/PowerUp/PowerUps/Jump/JumpBehaviour.cs
@@ -5,22 +5,6 @@ using UnityEngine.InputSystem;
 using System.Threading.Tasks;
 using System.Threading;
 
-public struct JumpFeature{
-    public float MaxJumpTime { get; }
-    public float MaxJumpHeight { get; }
-    public float TimeToApex { get; } 
-    public float JumpGravity { get; } 
-    public float IniJumpVelocity { get; }
-
-    public JumpFeature(float maxJumpTime, float maxJumpHeight){
-        MaxJumpTime = maxJumpTime;
-        MaxJumpHeight = maxJumpHeight;
-        TimeToApex = maxJumpTime / 2;
-        JumpGravity = (-2 * maxJumpHeight) / Mathf.Pow(TimeToApex, 2);
-        IniJumpVelocity = (2 * maxJumpHeight) / TimeToApex;
-    }
-}
-
 public class JumpBehaviour : MonoBehaviour{
 
     PlayerManager PlayerManager{ get{ return PlayerManager.instance; } }
