+++ Assets/Scripts/PowerUp/PowerUps/Jump/JumpFeature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f3d21fa3ebb34942ad7b23e14611bd6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
